apps/ai/requirements.txt
# Dependências principais (compatível com Python 3.12)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0  # Para análise estatística
pyarrow>=14.0.0  # Para suporte ao formato parquet
python-dotenv>=1.0.0

# Processamento de dados
matplotlib>=3.8.0
seaborn>=0.13.0

# Deep Learning (PyTorch) - latest stable
torch>=2.1.0
torchvision>=0.16.0

# API e Comunicação
flask>=3.0.0
requests>=2.31.0

# Banco de dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Ferramentas de desenvolvimento
pytest>=7.4.0
jupyter>=1.0.0
black>=23.12.0

# Logging e monitoramento
python-json-logger==2.0.7